requirements.txt
ultralytics>=8.3.0
PyQt6>=6.4.0
opencv-python>=4.7.0
numpy>=1.23.0
pillow>=9.4.0
fpdf2>=2.7.0
matplotlib>=3.7.0
